BackupManager.py: troubleshoot problems with NIM JESS settings
This problem arises and is noticeable when you have complex tuner setup, type JESS Unicable.

If we launch the order to unzip tar.gz in a Stage 1 primary level, the NIMS configuration can be lost in the configuration file, it must be taken to the Stage 6 level so that this does not happen, with this we make sure that the El configuration file remains intact.

I add fileExists so it doesn't duplicate the unzipped targ.gz and run twice.

This problem is difficult to capture, you must have this type of HW and antenna configuration to be able to reproduce it.