SoftcamManager.py: show buttons and actions only when available
Examples:
* red button + action:
If there is no cam in softcams it will only show the red button (close), it will not show the rest of the buttons and actions since there is no cam available.
* Green button + action:
If there is a cam, it will show a green button with action "start" if the cam has not started and action "stop" if it was already started.
* Yellow button + action:
Action Yellow button with "Restart" will only be shown if the cam is active (we cannot restart a cam if it is not previously activated).
If we select a cam that is not activated, it will not show a yellow button or its action.
* Blue button + action:
It will only be shown if there is a cam in softcams, if it does not exist it will not show a button or action, if on the contrary there is a cam the blue button will show "Disable startup" if the cam was selected in autostart and "Enable startup" if we did not have it selected in autostart.