Improve BackupManager, ImageManager and MountManager modules.
--BackupManager: Map indentation is corrected as well as conditions.
More correct lab texts are reviewed.
I level up loop to be able to have a "lab7" on the screen in case we have an OSError error on the device, then we are going to see said error on the screen.
The colored keys are only active if we have a device available, otherwise they will not be displayed or activated.
The actions menu is not accessed to avoid seeing incorrect elements.

--ImageManager: More correct "lab7" texts are adapted.
The color keys are only active if you have a device available to choose the color action.
The actions menu is not accessed to avoid seeing incorrect elements.

--MountManager: the colored keys are only active if we have a device available, otherwise they will not be shown or any execution will be activated.